more advanced haptics.  Now branches based on dominant hand and whether we're gripping or not

+++ HaloCEVR/WeaponHapticsConfig.h
@@ -6,7 +6,7 @@ using json = nlohmann::json;
 
 struct WeaponHapticArg
 {
-    int StartSecondsDelay;
+    float StartSecondsDelay;
     float DurationSeconds;
     float Frequency;
     float Amplitude;
